@@ -1,4 +1,4 @@
-package com.in28minutes.microservices.limitsservice;
+package com.in28minutes.microservices.limitsservice.configuration;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
@@ -6,24 +6,23 @@ import org.springframework.stereotype.Component;
 @Component
 @ConfigurationProperties("limits-service")
 public class Configuration {
-	
 	private int minimum;
 	private int maximum;
-
-	public void setMinimum(int minimum) {
-		this.minimum = minimum;
-	}
-
-	public void setMaximum(int maximum) {
-		this.maximum = maximum;
-	}
 
 	public int getMinimum() {
 		return minimum;
 	}
 
+	public void setMinimum(int minimum) {
+		this.minimum = minimum;
+	}
+
 	public int getMaximum() {
 		return maximum;
+	}
+
+	public void setMaximum(int maximum) {
+		this.maximum = maximum;
 	}
 
 }
